added Realtime Observer
